requirements.txt
asgiref==3.3.0
Django==3.1.3
django-bootstrap-breadcrumbs==0.9.2
django-bootstrap-form==3.4
django-crispy-forms==1.9.2
django-filter==2.4.0
django-widget-tweaks==1.4.8
PyMySQL==0.10.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
django-ckeditor
gunicorn
django-heroku

